\name{oct4}
\docType{data}
\alias{oct4}
\title{Oct4 data}
\description{
 This is a subset of the Oct4 data containing 12584 probes on chromosome
 20. The data were log2 transformed and quantile-normalized. 
}
\usage{data(oct4)}
\source{\url{http://jura.wi.mit.edu/young_public/hESregulation/Data_download.html}} 
\references{Boyer LA, Lee TI, Cole MF, et al. (2005).  Core
  transcriptional regulatory circuitry in human embryonic stem cells.  Cell,122(6), 828-30.}
\keyword{datasets}
