\name{p53}
\docType{data}
\alias{p53}
\title{p53 data}
\description{
 This is a subset of the p53 data containing 10000 probes on chromosome
 22. The data were log2 transformed and quantile-normalized. 
}
\usage{data(p53)}
\source{\url{http://www.gingeras.org/affy_archive_data/publication/tfbs/}}
\references{Cawley S, Bekiranov S, Ng HH, et al. (2004).  Unbiased
  mapping of transcription factor binding sites along human chromosomes
  21 and 22 points to widespread regulation of noncoding RNAs.  Cell, 116(4), 499-509.}
\keyword{datasets}
